Google and Facebook Authentications have been moved from SignUp to Login since it automaically logs you in with those methods.
