+++ src/CSS/SignUp.css
@@ -70,6 +70,19 @@ form > label > span {
     transition: 0.3s;
     margin-bottom: 1%;
   }
+
+  .sign__Up__Container > form > button {
+    background-color: #f7f7f7;
+    color: #333;
+    width: 100%;
+    padding: 7px 0;
+    border-radius: 20px;
+    border: none;
+    border-color: #a56f3f;
+    cursor: pointer;
+    transition: 0.3s;
+    margin-bottom: 1%;
+  }
   
   .social__SignUp {
     display: flex;
