+++ src/Components/Login.js
@@ -2,9 +2,13 @@ import React, { useState } from 'react'
 import './../CSS/Login.css'
 import Navbar from './Navbar'
 import SignUp from './SignUp'
-import { getAuth, signInWithEmailAndPassword } from "firebase/auth";
+import { getAuth, signInWithEmailAndPassword, createUserWithEmailAndPassword, signInWithPopup, GoogleAuthProvider, FacebookAuthProvider } from "firebase/auth";
 import { useNavigate } from 'react-router-dom';
 import { useDispatch } from 'react-redux';
+
+import GoogleIcon from '@mui/icons-material/Google';
+import FacebookOutlinedIcon from '@mui/icons-material/FacebookOutlined';
+
 
 function Login() {
 
@@ -42,6 +46,70 @@ function Login() {
       });
   }
 
+    // CREATES ACCOUNTS WITH GMAIL 
+    function handleGoogleSignIn() {
+      const provider = new GoogleAuthProvider()
+      const auth = getAuth()
+  
+      signInWithPopup(auth, provider)
+      .then((result) => {
+        // This gives you a Google Access Token. You can use it to access the Google API.
+        const credential = GoogleAuthProvider.credentialFromResult(result);
+        const token = credential.accessToken;
+        // The signed-in user info.
+        const user = result.user;
+        console.log(`User ${user.uid} signed in with Google`);
+        // IdP data available using getAdditionalUserInfo(result)
+        // ...
+      }).catch((error) => {
+        // Handle Errors here.
+        const errorCode = error.code;
+        const errorMessage = error.message;
+        // The email of the user's account used.
+        const email = error.customData.email;
+        // The AuthCredential type that was used.
+        const credential = GoogleAuthProvider.credentialFromError(error);
+        // ...
+        console.log(error.message);
+      });
+    }
+
+    function handleFacebookSignIn() {
+      const provider = new FacebookAuthProvider();
+      const auth = getAuth()
+  
+      signInWithPopup(auth, provider)
+      .then((result) => {
+        // The signed-in user info.
+        const user = result.user;
+  
+        // This gives you a Facebook Access Token. You can use it to access the Facebook API.
+        const credential = FacebookAuthProvider.credentialFromResult(result);
+        const accessToken = credential.accessToken;
+        console.log(`User ${user.uid} signed in with Google`);
+        // IdP data available using getAdditionalUserInfo(result)
+        // ...
+      })
+      .catch((error) => {
+        // Handle Errors here.
+        const errorCode = error.code;
+        const errorMessage = error.message;
+        // The email of the user's account used.
+        const email = error.customData.email;
+        // The AuthCredential type that was used.
+        const credential = FacebookAuthProvider.credentialFromError(error);
+        console.log(error.message);
+        // ...
+      });
+    }
+
+    // *******************************************************************
+    // *******************************************************************
+    // ******* GO BACK AND INSTALL FB AUTHENTICATION ON FIREBASE!!!! *****
+    // ************* DELETE THIS COMMENT WHEN YOURE DONE!!!! *************
+    // *******************************************************************
+    // *******************************************************************
+
 
   return (
     <>
@@ -72,6 +140,13 @@ function Login() {
                 <button className='signIn__button' onClick={handleSignIn}>Sign In</button>
                 <button className='forgotPass__button'>Forgot Password</button>
 
+                <h3>Or Sign in With:</h3>
+
+                <div className="social__SignUp">
+                  <GoogleIcon style={{ marginRight: '10px' }} onClick={handleGoogleSignIn} />
+                  <FacebookOutlinedIcon style={{ marginLeft: '10px' }} onClick={handleFacebookSignIn} />
+                </div>
+
                 <div className="create__account">
                   <p>Don't have an account?</p>
                   <a href="#" onClick={handleRegister}>Sign Up</a>
@@ -82,8 +157,8 @@ function Login() {
         ) : (
           <>
           <div className="signup__form">
-            <SignUp />
-            <button onClick={getLoginForm}>Login</button>
+            <SignUp getLoginForm={getLoginForm} />
+            {/* <button onClick={getLoginForm}>Login</button> */}
           </div>
           </>
         )}
