+++ src/Components/SignUp.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react'
 import './../CSS/SignUp.css'
 import { addUser } from '../Firebase/Firebase'
-import { getAuth, createUserWithEmailAndPassword, signInWithPopup, GoogleAuthProvider, FacebookAuthProvider } from "firebase/auth";
+import { getAuth, createUserWithEmailAndPassword } from "firebase/auth";
 import GoogleIcon from '@mui/icons-material/Google';
 import FacebookOutlinedIcon from '@mui/icons-material/FacebookOutlined';
 
@@ -13,7 +13,7 @@ function PasswordError({ errorMessage }) {
   );
 }
 
-function SignUp() {
+function SignUp({ getLoginForm }) {
 
   const [firstName, setFirstName] = useState('')
   const [lastName, setLastName] = useState('')
@@ -125,73 +125,6 @@ function SignUp() {
       });
   }
 
-  // CREATES ACCOUNTS WITH GMAIL 
-  function handleGoogleSignIn() {
-    const provider = new GoogleAuthProvider()
-    const auth = getAuth()
-
-    signInWithPopup(auth, provider)
-    .then((result) => {
-      // This gives you a Google Access Token. You can use it to access the Google API.
-      const credential = GoogleAuthProvider.credentialFromResult(result);
-      const token = credential.accessToken;
-      // The signed-in user info.
-      const user = result.user;
-      console.log(`User ${user.uid} signed in with Google`);
-      // IdP data available using getAdditionalUserInfo(result)
-      // ...
-    }).catch((error) => {
-      // Handle Errors here.
-      const errorCode = error.code;
-      const errorMessage = error.message;
-      // The email of the user's account used.
-      const email = error.customData.email;
-      // The AuthCredential type that was used.
-      const credential = GoogleAuthProvider.credentialFromError(error);
-      // ...
-      console.log(error.message);
-    });
-
-
-  }
-
-// *******************************************************************
-// *******************************************************************
-// ******* GO BACK AND INSTALL FB AUTHENTICATION ON FIREBASE!!!! *****
-// ************* DELETE THIS COMMENT WHEN YOURE DONE!!!! *************
-// *******************************************************************
-// *******************************************************************
-
-  function handleFacebookSignIn() {
-    const provider = new FacebookAuthProvider();
-    const auth = getAuth()
-
-    signInWithPopup(auth, provider)
-    .then((result) => {
-      // The signed-in user info.
-      const user = result.user;
-
-      // This gives you a Facebook Access Token. You can use it to access the Facebook API.
-      const credential = FacebookAuthProvider.credentialFromResult(result);
-      const accessToken = credential.accessToken;
-      console.log(`User ${user.uid} signed in with Google`);
-      // IdP data available using getAdditionalUserInfo(result)
-      // ...
-    })
-    .catch((error) => {
-      // Handle Errors here.
-      const errorCode = error.code;
-      const errorMessage = error.message;
-      // The email of the user's account used.
-      const email = error.customData.email;
-      // The AuthCredential type that was used.
-      const credential = FacebookAuthProvider.credentialFromError(error);
-      console.log(error.message);
-      // ...
-    });
-
-  }
-
   return (
     <div className='sign__Up'>
       <div className="sign__Up__Container">
@@ -263,14 +196,8 @@ function SignUp() {
             <p className={hasSpecialChar ? 'green' : 'red'}>Must contain a special character: !@#$%^&*()</p>
           </div>
           <button type='submit'>Sign Up</button>
+          <button onClick={getLoginForm}>Return to Login</button>
         </form>
-        <h3>Or Sign in With:</h3>
-
-        <div className="social__SignUp">
-          <GoogleIcon style={{ marginRight: '10px' }} onClick={handleGoogleSignIn} />
-          <FacebookOutlinedIcon style={{ marginLeft: '10px' }} onClick={handleFacebookSignIn} />
-        </div>
-
       </div>
     </div>
   )
